remove AsyncHelper in DynamicEntityPropertiesTestBase
