+++ test/Abp.Zero.SampleApp.Tests/DynamicEntityProperties/DynamicEntityPropertiesTestBase.cs
@@ -1,13 +1,8 @@
 ﻿using System;
 using System.Linq;
-using System.Threading;
 using System.Threading.Tasks;
-using Abp.Authorization;
 using Abp.DynamicEntityProperties;
-using Abp.Localization;
-using Abp.Modules;
 using Abp.Threading;
-using Abp.UI.Inputs;
 using Abp.Zero.SampleApp.EntityHistory;
 using Castle.MicroKernel.Registration;
 using Microsoft.AspNet.Identity;
@@ -42,8 +37,9 @@ namespace Abp.Zero.SampleApp.Tests.DynamicEntityProperties
 
         protected void RunAndCheckIfPermissionControlled(Action function, string requiredPermission = TestPermission)
         {
-            var user = AsyncHelper.RunSync(() => UserManager.FindByIdAsync(AbpSession.UserId.Value));
-            AsyncHelper.RunSync(() => ProhibitPermissionAsync(user, requiredPermission));
+            var user = UserManager.FindById(AbpSession.UserId.Value);
+            var permission = PermissionManager.GetPermission(requiredPermission);
+            ProhibitPermission(user, permission);
 
             bool isExceptionThrown = false;
             try
@@ -148,39 +144,6 @@ namespace Abp.Zero.SampleApp.Tests.DynamicEntityProperties
             var substitute = Substitute.For<T>();
             LocalIocManager.IocContainer.Register(Component.For<T>().Instance(substitute).IsDefault());
             return substitute;
-        }
-    }
-
-    public class MyDynamicEntityPropertyDefinitionProvider : DynamicEntityPropertyDefinitionProvider
-    {
-        public override void SetDynamicEntityProperties(IDynamicEntityPropertyDefinitionContext context)
-        {
-            context.Manager.AddAllowedInputType<SingleLineStringInputType>();
-            context.Manager.AddAllowedInputType<CheckboxInputType>();
-            context.Manager.AddAllowedInputType<ComboboxInputType>();
-
-            context.Manager.AddEntity<Country, int>();
-        }
-    }
-
-    public class DynamicEntityPropertiesTestAuthorizationProvider : AuthorizationProvider
-    {
-        public override void SetPermissions(IPermissionDefinitionContext context)
-        {
-            context.CreatePermission(DynamicEntityPropertiesTestBase.TestPermission,
-                new FixedLocalizableString(DynamicEntityPropertiesTestBase.TestPermission));
-        }
-    }
-
-    [DependsOn(typeof(SampleAppTestModule))]
-    public class DynamicEntityPropertiesTestModule : AbpModule
-    {
-        public override void PreInitialize()
-        {
-            IocManager.RegisterAssemblyByConvention(typeof(DynamicEntityPropertiesTestModule).Assembly);
-
-            Configuration.Authorization.Providers.Add<DynamicEntityPropertiesTestAuthorizationProvider>();
-            Configuration.DynamicEntityProperties.Providers.Add<MyDynamicEntityPropertyDefinitionProvider>();
         }
     }
 }
